Begin rewrite of engine
