+++ ConsoleGameEngine/old/CGEMath.cs
