+++ ConsoleGameEngine/old/StorageTypes.cs
