+++ ConsoleGameEngine/old/WindowBuilder.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ConsoleGameEngine.old
+{
+    public class WindowBuilder
+    {
+        #region Consts
+        //-============================-
+        //- BORDER CHARACTER CONSTANTS -
+        //-============================-
+        public const char charHorizontalBorder = '\x2550';
+        public const char charVerticalBorder = '\x2551';
+        public const char charBottomLeftCornerBorder = '\x255A';
+        public const char charTopLeftCornerBorder = '\x2554';
+        public const char charBottomRightCornerBorder = '\x255D';
+        public const char charTopRightCornerBorder = '\x2557';
+        public const char charNotTopTBorder = '\x2566';
+        public const char charNotLeftTBorder = '\x2560';
+        public const char charNotBottomTBorder = '\x2569';
+        public const char charNotRightTBorder = '\x2563';
+        public const char charXBorder = '\x256C';
+        #endregion
+
+        //public WindowBuilder(uint width, uint height)
+        //{
+
+        //}
+    }
+}
